fix: jsonrpc to support both positional parameters and named parameters (#16)
https://github.com/notional-labs/subnode/issues/15

+++ config_test.go
@@ -24,7 +24,7 @@ upstream:
 func TestGetBackendNodeType(t *testing.T) {
 	b := Backend{
 		Rpc:    "http://pruned",
-		Blocks: []int{10},
+		Blocks: []int64{10},
 	}
 
 	assert.Equal(t, GetBackendNodeType(&b), BackendNodeTypePruned)
